upload pause state before unloading a task
